fix import, pep8 and refactor
