+++ geotrek/common/apps.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from django.utils.translation import ugettext_lazy as _
 
-from geotrek.core.appconfig import GeotrekConfig
+from geotrek.appconfig import GeotrekConfig
 
 
 class CommonConfig(GeotrekConfig):
